requirements.txt
transformers
torchensemble
timm
pillow
streamlit